% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{fscca}
\alias{fscca}
\title{Fast sparse canonical correlation}
\usage{
fscca(X, Y, penalty_x, penalty_y, lam_x, lam_y, k_folds = 5L,
  n_components = 1L, center = TRUE, scale = FALSE)
}
\arguments{
\item{X}{a matrix of dimension n x p}

\item{Y}{a matrix of dimension n x q}

\item{penalty_x}{a string indicating the penalty function to use on matrix
X. Currently only "lasso" is implemented.}

\item{penalty_y}{a string indicating the penalty function to use on matrix
Y. Currently only "lasso" is implemented.}

\item{lam_x}{a numeric vector of tuning parameters on X}

\item{lam_y}{a numeric vector of tuning parameters on Y}

\item{k_folds}{an integer denoting the number of folds to use in cros-validation}

\item{n_components}{the number of components to compute}

\item{center}{center the columns to mean zero?}

\item{scale}{scale the columns to standard deviation one?}
}
\value{
A list with matrices:
\describe{
    \item{A}{ A matrix of dimension p x n_components of the canonical
vector }
    \item{B}{ A matrix of dimension q x n_components of the canonical
vector }
    \item{U}{ A matrix of dimension n x n_components of the X * a }
    \item{V}{ A matrix of dimension n x n_components of the Y * b }
    \item{lambda}{ A matrix of dimension n_components x n of the optimal
tuning parameters}
    \item{covar}{ The average cross-validated covariance }
}
}
\description{
This is a fast implementation of sparse canonical correlation analysis.
}

