% Generated by roxygen2 (4.0.2): do not edit by hand
\name{sparse_nipals}
\alias{sparse_nipals}
\title{Sparse NIPALS CCA algorithm}
\usage{
sparse_nipals(Xr, Yr, lamx, lamy)
}
\arguments{
\item{lamx}{a positive enalty on 'a'}

\item{lamy}{a positive penalty on 'b'}

\item{x}{a matrix x that has been centered and scaled}

\item{y}{a matrix y that has been centered and scaled}
}
\value{
a list containing a1 and b1
}
\description{
Sparse NIPALS CCA algorithm
}

