% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nipals}
\alias{nipals}
\title{NIPALS CCA algorithm}
\usage{
nipals(Xr, Yr)
}
\arguments{
\item{x}{a matrix x that has been centered and scaled}

\item{y}{a matrix y that has been centered and scaled}
}
\value{
a list containing a1 and b1
}
\description{
NIPALS CCA algorithm
}

