% Generated by roxygen2 (4.0.2): do not edit by hand
\name{nipals}
\alias{nipals}
\title{NIPALS CCA algorithm}
\usage{
nipals(X, Y)
}
\arguments{
\item{X}{a matrix X that has been centered and scaled}

\item{Y}{a matrix Y that has been centered and scaled}
}
\value{
a list containing a, b, u, v, and rho
}
\description{
NIPALS CCA algorithm
}

