% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{sparse_nipals}
\alias{sparse_nipals}
\title{Sparse NIPALS CCA algorithm}
\usage{
sparse_nipals(X, Y, penalty_x, penalty_y, lamx, lamy)
}
\arguments{
\item{X}{a matrix X that has been centered and scaled}

\item{Y}{a matrix Y that has been centered and scaled}

\item{penalty_x}{A character string of type "lasso"}

\item{penalty_y}{A character string of type "lasso"}

\item{lamx}{a positive penalty on 'a'}

\item{lamy}{a positive penalty on 'b'}
}
\value{
a list containing a, b, u, v, and rho (covariance)
}
\description{
Sparse NIPALS CCA algorithm
}

